+++ examples/REST/server.php
@@ -2,11 +2,11 @@
 
 include __DIR__ . '../../../vendor/autoload.php';
 
-use Rubix\Server\RESTServer;
-use Rubix\Server\Http\Middleware\ResponseTime;
 use Rubix\ML\Datasets\Generators\Blob;
 use Rubix\ML\Datasets\Generators\Agglomerate;
 use Rubix\ML\Classifiers\KNearestNeighbors;
+use Rubix\Server\RESTServer;
+use Rubix\Server\Http\Middleware\AccessLogGenerator;
 use Rubix\ML\Other\Loggers\Screen;
 
 $generator = new Agglomerate([
@@ -15,14 +15,14 @@ $generator = new Agglomerate([
     'blue' => new Blob([0, 0, 255], 10.0),
 ]);
 
-$estimator = new KNearestNeighbors(3);
+$estimator = new KNearestNeighbors(5);
 
 $dataset = $generator->generate(100);
 
 $estimator->train($dataset);
 
 $server = new RESTServer('127.0.0.1', 8080, null, [
-    new ResponseTime(),
+    new AccessLogGenerator(new Screen()),
 ]);
 
 $server->setLogger(new Screen());