+++ examples/RPC/client.php
@@ -3,22 +3,25 @@
 include __DIR__ . '../../../vendor/autoload.php';
 
 use Rubix\Server\RPCClient;
-use Rubix\ML\Datasets\Unlabeled;
+use Rubix\ML\Datasets\Generators\Blob;
+use Rubix\ML\Datasets\Generators\Agglomerate;
 
 $client = new RPCClient('127.0.0.1', 8888, false, [
     'Authorization' => 'Basic ' . base64_encode('test:secret'),
 ]);
 
-$dataset = new Unlabeled([
-    [228, 28, 138],
-    [44, 129, 208],
-    [27, 165, 7],
+$generator = new Agglomerate([
+    'red' => new Blob([255, 0, 0], 10.0),
+    'green' => new Blob([0, 128, 0], 10.0),
+    'blue' => new Blob([0, 0, 255], 10.0),
 ]);
 
-print_r($client->predict($dataset));
+$dataset = $generator->generate(10)->randomize();
 
-print_r($client->predictSample($dataset->sample(1)));
+$predictions = $client->predict($dataset);
 
-print_r($client->proba($dataset));
+$probabilities = $client->proba($dataset);
 
-print_r($client->probaSample($dataset->sample(2)));
+print_r($predictions);
+
+print_r($probabilities);