+++ src/Commands/ScoreSample.php
@@ -16,7 +16,7 @@ class ScoreSample extends Command
     /**
      * The sample to rank.
      *
-     * @var mixed[]
+     * @var (string|int|float)[]
      */
     protected $sample;
 
@@ -37,7 +37,7 @@ class ScoreSample extends Command
     }
 
     /**
-     * @param mixed[] $sample
+     * @param (string|int|float)[] $sample
      * @throws \Rubix\Server\Exceptions\ValidationException
      */
     public function __construct(array $sample)
@@ -52,7 +52,7 @@ class ScoreSample extends Command
     /**
      * Return the sample to predict.
      *
-     * @return mixed[]
+     * @return (string|int|float)[]
      */
     public function sample() : array
     {