+++ src/Http/Middleware/AccessLogGenerator.php
@@ -7,16 +7,18 @@ use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Log\LoggerInterface;
 
 /**
- * Access Log
+ * Access Log Generator
  *
- * Generates an HTTP access log similar to the Apache log format.
+ * Generates an HTTP access log using a format similar to the Apache log format.
  *
  * @category    Machine Learning
  * @package     Rubix/Server
  * @author      Andrew DalPino
  */
-class AccessLog implements Middleware
+class AccessLogGenerator implements Middleware
 {
+    public const UNKNOWN = '-';
+
     /**
      * A PSR-3 logger instance.
      *
@@ -45,25 +47,31 @@ class AccessLog implements Middleware
 
         $server = $request->getServerParams();
 
-        $ip = $server['REMOTE_ADDR'] ?? '-';
+        $ip = $server['REMOTE_ADDR'] ?? self::UNKNOWN;
 
         $method = $request->getMethod();
 
-        $uri = $request->getUri()->getPath();
+        $uri = $request->getUri();
 
-        $version = 'HTTP/' . $request->getProtocolVersion();
+        $version = "HTTP/{$request->getProtocolVersion()}";
+
+        $requestString = "\"$method {$uri->getPath()} $version\"";
 
         $status = $response->getStatusCode();
 
         $size = $response->getBody()->getSize();
 
-        $headers = $request->getHeaders();
+        $referrer = $request->hasHeader('Referer')
+            ? "\"{$request->getHeaderLine('Referer')}\""
+            : self::UNKNOWN;
 
-        $agent = $headers['User-Agent'][0] ?? '-';
+        $agent = $request->hasHeader('User-Agent')
+            ? "\"{$request->getHeaderLine('User-Agent')}\""
+            : self::UNKNOWN;
 
-        $record = "$ip '$method $uri $version' $status $size $agent";
+        $entry = "$ip $requestString $status $size $referrer $agent";
 
-        $this->logger->info($record);
+        $this->logger->info($entry);
 
         return $response;
     }